refactor: move generic accumulator to multiexp/curve package ( PROOF-589 )
* require accumulator to satisfy the Element concept
